Add Dockerfile to generate a base image for plugins (#12)
Add Dockerfile for generating base image for plugin

This commit backports most of the work originally done in `slicer_cli_web_plugin`
project so that the created Docker image can serve as a generic base for 
other Slicer CLI web plugins (either python or C++).

In a nutshell, this `Dockerfile` does the following:

* `apt-get install` system pre-requisites
* Download and Install miniconda
* Download and Install CMake
* git clone and build ITK 4.10
* git clone and build SlicerExecutionModel
* Install ctk_cli=1.3.1 from my conda channel
* Copy in files of slicer_cli_web


It also refactors the code to remove the dependency on Girder. The only change
required was to remove the use of ``girder.logger.Exception``.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>

--
* Add Dockerfile to create base image

* Add .dockerignore

* Fix error in .dockerignore

* Removing girder.logger.Exception to avoid installing girder in Docker

This is only used as an entrypoint for all docker plugins of slicer_cli_web and HistomicsTK. Using girder.logger.Exception will require us to install girder. I am removing this to avoid installing it as girder would not be used for anything else

* Fix bug in cli_list_entrypoint.py

* Change build_path to BUILD_PATH