Pin to version 2.3.6 of bootstrap-colorpicker
The most recent release (2.4.0) broke compatibility with our
webpack build system. Namely, the factory it uses is intending
to see if the "exports" symbol is globally defined, but it is
doing so using "if (typeof exports === "object")". It's actually
picking up the local version of the "exports" variable that is
an internal detail of webpack.

If they release another version after 2.4.0, we can try to upgrade
to that to see if they've resolved this problem.