Fix parsing JSON lists of docker images.
When docker images were passed as a JSON list, they had to have BOTH a version AND a digest (which I don't think is allowed), rather than EITHER a version OR a digest.  Now, simple string and lists are validated with the same code.

Made some of the formatting more consistent.