Support partially valid CLIs.
If a CLI can return xml for its commands, but some of those commands cannot be made into valid endpoints, still generate endpoints for valid commands.

Otherwise, the failure to generate an endpoint prevents ANY cli in the system from being listed.

Make some of the formatting more consistent and clean up some text.

For testing, update the small docker image with an example that is missing a default.