When creating a input spec, don't fetch the parent.
This was done when we first started asking for images as files rather than items.  We have the ability to ask for items distinct from files, so if something needs a whole item, it really should ask for it.  Otherwise, for a multi-file item, we end up downloading all of the files, even though only one is asked for.