Use the fileId, not originalId for large images.
Before, the CLI was passed the original image, not the large image.  This was a hold over from before CLIs used large image.

large_image specific code should be pulled out of slicer_cli_web and moved to large_image, but until that is done, this fixes the current issue.