Add a new event `h:analysis:rendered`
Sometimes, and I'm at a loss for why, the rendering of the panels occur
asynchronously.  The use case for this event in HistomicsTK requires
that default values be inserted *after* the rendering occurs.  This adds
a new event that HistomicsTK can listen to that is guaranteed to occur
after the render.