Support optional file outputs.
Prior to this, optional file outputs were ignored.  This marks a file
output with required = (boolean) based on whether is is optional
(doesn't have an index).

If it is optional, an empty value will validate within the javascript.

This also fixes a Girder warning about using path vs url in restRequest.