Squelch non-useful warnings.
When the slicer_cli specifications are parsed, they weren't using the
channel to reduce warnings.  As such, many errors were printed that
shouldn't be displayed.

Also, make a few other changes to reduce warnings where they are not
useful.  Ideally, after this PR, all emitted warnings are actual issues
that should be fixed.