Support defaultNameMatch and defaultPathMatch attributes.
On some input types (file, item, image, folder), you can specify a
defaultNameMatch and defaultPathMatch attributes (much like
fileExtension is used on some outputs).  These are hints to the UI to
find a matching file.  For the Girder UI, this is done on the resource
path, with the name match applying to the specific Girder resource and
the path match applying to the full resource path (e.g., /user/<user
name>/<folder name>.../<resource name>).

These are regex; if you need to match the whole string, add begin and
end anchors (e.g., '\.png$' will match any .png, '^a.*\.png$' will match
a png that starts with 'a').